connected-contexts/pyproject.toml
[project]
name = "connected-contexts"
version = "0.1.0"
description = "Influence Tensor Experiment Pipeline for studying contextual effects on model influence"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.30.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.33.0",
    "plotly>=5.18.0",
    "scikit-learn>=1.3.0",
]

[project.scripts]
generate-contexts = "scripts.generate_contexts:main"
review-contexts = "scripts.review_contexts:main"
assemble-queries = "scripts.assemble_queries:main"
run-all-queries = "scripts.run_all_queries:main"
assemble-csvs = "analysis.assemble_csvs:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts", "analysis", "visualization"]
